+++ exonum/src/events/noise/tests.rs
@@ -17,7 +17,7 @@ use snow::wrappers::crypto_wrapper::Dh25519;
 use snow::NoiseBuilder;
 
 use crypto::PUBLIC_KEY_LENGTH;
-use crypto::{gen_keypair, into_x25519_keypair};
+use crypto::{gen_keypair, x25519::into_x25519_keypair};
 
 #[test]
 fn test_convert_ed_to_curve_dh() {
@@ -31,12 +31,12 @@ fn test_convert_ed_to_curve_dh() {
 
     // Do DH.
     let mut keypair_i: Dh25519 = Default::default();
-    keypair_i.set(&secret_key_i[..PUBLIC_KEY_LENGTH]);
+    keypair_i.set(secret_key_i.as_ref());
     let mut output_i = [0u8; PUBLIC_KEY_LENGTH];
     keypair_i.dh(public_key_r.as_ref(), &mut output_i);
 
     let mut keypair_r: Dh25519 = Default::default();
-    keypair_r.set(&secret_key_r[..PUBLIC_KEY_LENGTH]);
+    keypair_r.set(secret_key_r.as_ref());
     let mut output_r = [0u8; PUBLIC_KEY_LENGTH];
     keypair_r.dh(public_key_i.as_ref(), &mut output_r);
 
@@ -58,13 +58,13 @@ fn test_converted_keys_handshake() {
     let (public_key_r, secret_key_r) = into_x25519_keypair(public_key_r, secret_key_r).unwrap();
 
     let mut h_i = NoiseBuilder::new(PATTERN.parse().unwrap())
-        .local_private_key(&secret_key_i[..PUBLIC_KEY_LENGTH])
-        .remote_public_key(&public_key_r[..])
+        .local_private_key(secret_key_i.as_ref())
+        .remote_public_key(public_key_r.as_ref())
         .build_initiator()
         .expect("Unable to create initiator");
 
     let mut h_r = NoiseBuilder::new(PATTERN.parse().unwrap())
-        .local_private_key(&secret_key_r[..PUBLIC_KEY_LENGTH])
+        .local_private_key(secret_key_r.as_ref())
         .build_responder()
         .expect("Unable to create responder");